Split DuckDbFactCollector into per-domain partials (Lite parity)
Behavior-neutral partial-class split mirroring the Dashboard
SqlServerFactCollector split. The 28 Collect*FactsAsync methods are
fanned into six per-domain partial files (Waits, Config, Resources,
Storage, QueryPerf, Activity); the ctor, fields, CollectFactsAsync
orchestrator, and the ToInt64 helper stay in the base file. No call-site
or behavior changes — same class, same namespace.

Build: Build succeeded, 0 errors (PerformanceMonitor.sln, Debug).
Tests: Passed! 545 passed, 0 failed, 0 skipped (Lite.Tests).
